Move yugipedia module to client/yugipedia
