+++ tools/src/main/java/io/github/ygojson/tools/client/yugipedia/YugipediaApi.java
@@ -1,4 +1,4 @@
-package io.github.ygojson.tools.yugipedia.api;
+package io.github.ygojson.tools.client.yugipedia;
 
 import java.time.ZonedDateTime;
 
@@ -6,10 +6,10 @@ import retrofit2.Call;
 import retrofit2.http.GET;
 import retrofit2.http.Query;
 
-import io.github.ygojson.tools.yugipedia.api.params.Category;
-import io.github.ygojson.tools.yugipedia.api.params.PipeSeparated;
-import io.github.ygojson.tools.yugipedia.api.params.SortDirection;
-import io.github.ygojson.tools.yugipedia.api.response.QueryResponse;
+import io.github.ygojson.tools.client.yugipedia.params.Category;
+import io.github.ygojson.tools.client.yugipedia.params.PipeSeparated;
+import io.github.ygojson.tools.client.yugipedia.params.SortDirection;
+import io.github.ygojson.tools.client.yugipedia.response.QueryResponse;
 
 /**
  * Represents the Yugipedia API calls used by YGOJSON.
@@ -26,12 +26,12 @@ public interface YugipediaApi {
 	 */
 	@GET(
 		"api.php?action=query" +
-		"&format=json&formatversion=2" +
-		"&redirects=true" +
-		"&prop=revisions" +
-		"&rvprop=content|timestamp" +
-		"&generator=categorymembers" +
-		"&gcmsort=timestamp"
+			"&format=json&formatversion=2" +
+			"&redirects=true" +
+			"&prop=revisions" +
+			"&rvprop=content|timestamp" +
+			"&generator=categorymembers" +
+			"&gcmsort=timestamp"
 	)
 	public Call<QueryResponse> queryCategoryMembersByTimestamp(
 		@Query("gcmtitle") Category category,
@@ -50,14 +50,14 @@ public interface YugipediaApi {
 	 */
 	@GET(
 		"api.php?action=query" +
-		"&format=json&formatversion=2" +
-		"&redirects=true" +
-		"&prop=revisions|categories" +
-		"&rvprop=content|timestamp" +
-		"&generator=recentchanges" +
-		"&grctype=new|edit|categorize" +
-		"&grctoponly=true" +
-		"&cllimit=max"
+			"&format=json&formatversion=2" +
+			"&redirects=true" +
+			"&prop=revisions|categories" +
+			"&rvprop=content|timestamp" +
+			"&generator=recentchanges" +
+			"&grctype=new|edit|categorize" +
+			"&grctoponly=true" +
+			"&cllimit=max"
 	)
 	public Call<QueryResponse> queryRecentChanges(
 		@Query("grclimit") Integer resultsPerQuery,
@@ -73,10 +73,10 @@ public interface YugipediaApi {
 	 */
 	@GET(
 		"api.php?action=query" +
-		"&format=json&formatversion=2" +
-		"&redirects=true" +
-		"&prop=revisions" +
-		"&rvprop=content|timestamp"
+			"&format=json&formatversion=2" +
+			"&redirects=true" +
+			"&prop=revisions" +
+			"&rvprop=content|timestamp"
 	)
 	public Call<QueryResponse> queryPagesByTitle(
 		@Query("titles") PipeSeparated titles
