+++ tools/src/main/java/io/github/ygojson/tools/client/yugipedia/response/Page.java
@@ -1,4 +1,4 @@
-package io.github.ygojson.tools.yugipedia.api.response;
+package io.github.ygojson.tools.client.yugipedia.response;
 
 import java.util.List;
